Added Interaction System + Door Test
Can now create interactions by specifying a callback that will be called when player presses "e" while in range. As a test there is now a door to the right of spawn which only opens if the player is holding a teapot. Also, fixed some minor issues with the item system and added a function for removing items from the scene.

Interaction system currently shares a lot of code with the item system, so I will make it so that the item sys simply calls the interaction sys in the future.