Added bumped normals/changed material values
The look of the walls and how dark they are really depend on the material values(color, ambient, diffusivity...). Slightly changing a setting completely changes the mood of the scene. I still have not figured out how to make the texture not bright but still show up under the flashlight. I have fog that darkens things that are far away exponentially, but it also makes things look gray under the flashlight. Also, up close with no light, the textures are still very bright. Only if there was a way to not tie the texture value to the ambient light...