Gate Room
Added a simple room with a gate and a pressure plate. A box in the room can be picked up and placed on the pressure plate to open up the gate. Behind the gate is a key.

NOTE: For some reason applying a rotation to gate_room_base_transform breaks some wall collision. So only translations will work.