Added Jail Room
Located to the right of the first room.

Puzzle:
There are 4 cells in the room each with a gate in front of them. From left to right the cells contain: A wooden wall, the key, nothing, and a cube (meant to be an axe or some sort of tool to break the wooden wall). The lever in the middle of the room controls which cell is open. However, the cell containing the key is blocked by barrels and cannot be accessed directly.

Solution:
The cells are numbered 1-4 from left to right. Open cell 4 and pick up the cube (axe?). Open cell 1 and break the wooden wall. This allows you to access the key in cell 2.